backend/requirements.txt
# Flask Web框架
Flask==3.0.0
Flask-CORS==4.0.0

# HTTP请求库
requests==2.31.0

# 环境变量管理
python-dotenv==1.0.0

# 日志和工具
Werkzeug==3.0.1

# 可选：生产环境WSGI服务器
gunicorn==21.2.0

# 可选：阿里云SDK（如果需要使用官方SDK）
# dashscope==1.14.0

